refactor(core): rename to BuiltinInjection and remove CoreInjection
Remove CoreInjectionProxy and CoreInjectionProxyFactory placeholders.
MinimalWorld.GetInjectionProxyFactory now throws NotImplementedException.
Rename InjectorInjection types to BuiltinInjectionProxy and Factory.

Test MinimalWorld stubs override GetInjectionProxyFactory with the built-in factory.
